Moved a couple of version control interfaces out of the public API and into internal API. These were not reachable from any other public API and we're not planning to make the VCS integration extensible at this point (certainly something we want to do in the future I think, once source dependencies work well with Git).
