+++ subprojects/core-api/src/main/java/org/gradle/vcs/internal/VersionRef.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.vcs;
+package org.gradle.vcs.internal;
 
 import org.gradle.api.Incubating;
