+++ subprojects/version-control/src/main/java/org/gradle/vcs/internal/DefaultVersionRef.java
@@ -16,8 +16,6 @@
 
 package org.gradle.vcs.internal;
 
-import org.gradle.vcs.VersionRef;
-
 public class DefaultVersionRef implements VersionRef {
     @Override
     public String getVersion() {
